Create path when downloading LAS
Downloading a LAS from an URL was not creating the output path,
resulting in an error if the path does not yet exist. This bug was
always present, but not very exposed as the default cache location is
`/tmp`, which should exist. Commit 548b44b however changed the code in
`DataSources` to use the `LAS` constructor for downloading to the
default cache location, which now exposes everyone newly using the
`gettiles` functionality to this bug.